Lecture 5 Homework (Programming#1)
The following program prompts the user for a number of items to be packaged.•Each box can hold 10 items. Any left over items require an extra box. •The first 6 boxes cost $8 each and any boxes above the first 6, cost $5 each. •The program executes as shown in the example outputs below. •Design the functions needed to write this program and write the main code for this program.